+++ src/main/java/org/folio/marccat/dao/persistence/ClassificationType.java
@@ -18,15 +18,15 @@ public class ClassificationType extends T_SINGLE {
     return s == 12;
   }
 
-  public static boolean isLC(short s) {
+  public static boolean isLC(int s) {
     return s == 1;
   }
 
-  public static boolean isNLM(short s) {
+  public static boolean isNLM(int s) {
     return s == 6;
   }
 
-  public static boolean isNAL(short s) {
+  public static boolean isNAL(int s) {
     return s == 8;
   }
 }
